# coreApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Johnny Cms Api",
        "description": "Johnny Cms Api ",
        "contact": {
            "email": "johnny_login@163.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/getProduct": {
            "post": {
                "tags": [
                    "Cms 商品"
                ],
                "summary": "查询商品列表",
                "description": "查询商品列表",
                "operationId": "Module",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "查询页",
                        "required": true,
                        "schema": {
                            "type": "Integer"
                        }
                    },
                    {
                        "name": "size",
                        "in": "path",
                        "description": "查询每页显示数",
                        "required": true,
                        "schema": {
                            "type": "Integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "path",
                        "description": "搜索名称对象",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sid",
                        "in": "path",
                        "description": "查询所属分类",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/newProduct": {
            "post": {
                "tags": [
                    "Cms 商品"
                ],
                "summary": "创建商品信息",
                "description": "创建商品信息",
                "operationId": "Module",
                "parameters": [
                    {
                        "name": "sid",
                        "in": "path",
                        "description": "所属ID",
                        "required": true,
                        "schema": {
                            "type": "Integer"
                        }
                    },
                    {
                        "name": "proName",
                        "in": "path",
                        "description": "商品名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "proPrice",
                        "in": "path",
                        "description": "商品价格",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "proUnit",
                        "in": "path",
                        "description": "商品单位",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "proTitle",
                        "in": "path",
                        "description": "商品简介",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "proVal",
                        "in": "path",
                        "description": "商品详情",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/setProduct": {
            "post": {
                "tags": [
                    "Cms 商品"
                ],
                "summary": "更新商品信息",
                "description": "更新商品信息",
                "operationId": "Module",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "目标ID",
                        "required": true,
                        "schema": {
                            "type": "Integer"
                        }
                    },
                    {
                        "name": "sid",
                        "in": "path",
                        "description": "所属ID",
                        "required": true,
                        "schema": {
                            "type": "Integer"
                        }
                    },
                    {
                        "name": "proName",
                        "in": "path",
                        "description": "商品名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "proPrice",
                        "in": "path",
                        "description": "商品价格",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "proUnit",
                        "in": "path",
                        "description": "商品单位",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "proTitle",
                        "in": "path",
                        "description": "商品简介",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "proVal",
                        "in": "path",
                        "description": "商品详情",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/delProduct": {
            "post": {
                "tags": [
                    "Cms 商品"
                ],
                "summary": "删除商品信息",
                "description": "删除商品信息",
                "operationId": "Module",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "目标ID",
                        "required": true,
                        "schema": {
                            "type": "Integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    }
}